remove 'enhanced' label
